update fee to be uint256 and milli

+++ contracts/lib/Structs.sol
@@ -151,7 +151,7 @@ struct RefinanceFullfillment {
 //////////////////////////////////////////////////*/
 
 struct Fee {
-    uint16 rate;
+    uint256 rate;
     address recipient;
 }
